Implement interest and cash balance (#35)

+++ GhostfolioSidekick/Ghostfolio/API/MarketData.cs
@@ -1,11 +1,11 @@
 ﻿namespace GhostfolioSidekick.Ghostfolio.API
 {
-    public class MarketData
-    {
-        public DateTime Date { get; set; }  
+	public class MarketData
+	{
+		public DateTime Date { get; set; }
 
-        public string Symbol { get; set; }
+		public string Symbol { get; set; }
 
-        public double MarketPrice { get; set; }
-    }
+		public decimal MarketPrice { get; set; }
+	}
 }
